+++ ReleaseNotes.md
@@ -16,6 +16,7 @@
 - Added Quest and QuestManager system
 - Added quest events (OnQuestEvent trigger)
 - Added Quest actions (Give, Remove and Fail quest)
+- Added Token actions (Give, Remove)
 - Fixed several UI bugs
 
 ## V1.18.2